fix: 分离公共逻辑
